Change declaration retrieval to in-source matching
